Set up new layout structure to differentiate the legacy and new layout
