+++ code/ui/manager/src/components/layoutNew/Layout.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const Layout = () => {
+  return <div>New Layout</div>;
+};
